changed userId type to long

+++ src/main/java/com/banti/wallet/ums/repository/UserRepository.java
@@ -1,9 +1,8 @@
 package com.banti.wallet.ums.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-
 import com.banti.wallet.ums.model.User;
 
-public interface UserRepository extends JpaRepository<User,Integer>
-{ 
+public interface UserRepository extends JpaRepository<User,Long>
+{
 }
